Extract tag types to own directory
