+++ tenant-file/portal-app/src/pages/Tags-Page.tsx
@@ -1,40 +1,10 @@
 import React, { useState } from 'react';
 import EditTagRow from '../component/edit-tag-row';
 import AddTagRow from '../component/add-tag-row';
-import Tag from '../component/tag';
+import TagButton from '../component/tag';
 import StyledTagsDisplay from '../component/styles/TagsPage';
-
-const sampleData = [
-  {
-    name: 'Damages',
-    description: 'Windows, walls, appliance',
-    photoCount: 5,
-    id: 0,
-    color: '#00D084',
-  },
-  {
-    name: 'Conversations',
-    description: 'Records of texts or phone calls',
-    photoCount: 4,
-    id: 1,
-    color: '#0693E3',
-  },
-  {
-    name: 'Documents',
-    description: 'Lease, notices, etc.',
-    photoCount: 5,
-    id: 2,
-    color: '#FF6900',
-  },
-];
-
-type Tag = {
-  name: string;
-  description: string;
-  photoCount: number;
-  id: number;
-  color: string;
-};
+import { Tag } from '../types/tag';
+import { sampleData } from '../utility';
 
 const Tags = () => {
   // Sample data stored here. Can be swapped with redux or other method later.
@@ -61,7 +31,7 @@ const Tags = () => {
     return (
       <div className="flex-row" key={tag.id}>
         <div className="label">
-          <Tag tag={tag} />
+          <TagButton tag={tag} />
         </div>
         <div className="description">{tag.description}</div>
         <div className="photoCount">{tag.photoCount}</div>
