+++ tenant-file/portal-app/src/types/tag.ts
@@ -1,0 +1,15 @@
+export type Tag = {
+  name: string;
+  description: string;
+  photoCount: number;
+  id: number;
+  color: string;
+};
+
+export type Tags = {
+  name: string;
+  description: string;
+  photoCount: number;
+  id: number;
+  color: string;
+}[];
